hide user password field from response

+++ src/users/users.service.ts
@@ -1,9 +1,12 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { User } from './entities/user.entity';
 import { Repository } from 'typeorm';
+import { plainToClass } from 'class-transformer';
+import { UserDto } from './dto/user.dto';
+import { UpdateDto } from './dto/update.dto';
 
 @Injectable()
 export class UsersService {
@@ -14,11 +17,13 @@ export class UsersService {
   }
 
   async findAll(page, page_size) {
-    return this.usersRepository.find({ skip: page * page_size, take: page_size, select: ['id', 'full_name', 'created_at', 'updated_at', 'is_admin'] });
+    const users = await this.usersRepository.find({ skip: page * page_size, take: page_size });
+    return users.map(user => plainToClass(UserDto, user, { excludeExtraneousValues: true }))
   }
 
   async findOne(id: number) {
-    return this.usersRepository.findOneBy({id});
+    const user = await this.usersRepository.findOneBy({id});
+    return plainToClass(UserDto, user, { excludeExtraneousValues: true });
   }
 
   async findOneByUsername(username: string) {
@@ -29,17 +34,22 @@ export class UsersService {
   // Owner guard
   async update(id: number, updateUserDto: UpdateUserDto) {
     const user = await this.usersRepository.findOneBy({id});
-    
-    return this.usersRepository.save({...user, ...updateUserDto});
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+    const updatedUser = await this.usersRepository.save({...user, ...updateUserDto});
+    return plainToClass(UpdateDto, updatedUser, { excludeExtraneousValues: true });
   }
 
   // TO-DO
   // Owner guard
   async remove(id: number) {
     const user = await this.usersRepository.findOneBy({id});
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
 
-    // TO-DO
-    // Soft delete
-    return this.usersRepository.remove(user);
+    const removedUser = await this.usersRepository.remove(user);
+    return plainToClass(UserDto, removedUser, { excludeExtraneousValues: true });
   }
 }
